修复了一个会导致重定义的问题
