+++ pvzclass/Events/CoinCreateEvent.cpp
@@ -1,0 +1,26 @@
+#include "CoinCreateEvent.h"
+
+CoinCreateEvent::CoinCreateEvent()
+{
+	address = 0x40CCCE;
+}
+
+#if defined(_WIN64)
+void CoinCreateEvent::handle(CONTEXT& context)
+{
+	auto coin = std::make_shared<PVZ::Coin>(context.Rax);
+	for (int i = 0; i < listeners.size(); i++)
+	{
+		listeners[i](coin);
+	}
+}
+#else
+void CoinCreateEvent::handle(CONTEXT& context)
+{
+	auto coin = std::make_shared<PVZ::Coin>(context.Eax);
+	for (int i = 0; i < listeners.size(); i++)
+	{
+		listeners[i](coin);
+	}
+}
+#endif
