+++ pvzclass/Events/CoinCreateEvent.h
@@ -11,28 +11,3 @@ public:
 	CoinCreateEvent();
 	void handle(CONTEXT& context) override;
 };
-
-CoinCreateEvent::CoinCreateEvent()
-{
-	address = 0x40CCCE;
-}
-
-#if defined(_WIN64)
-void CoinCreateEvent::handle(CONTEXT& context)
-{
-	auto coin = std::make_shared<PVZ::Coin>(context.Rax);
-	for (int i = 0; i < listeners.size(); i++)
-	{
-		listeners[i](coin);
-	}
-}
-#else
-void CoinCreateEvent::handle(CONTEXT& context)
-{
-	auto coin = std::make_shared<PVZ::Coin>(context.Eax);
-	for (int i = 0; i < listeners.size(); i++)
-	{
-		listeners[i](coin);
-	}
-}
-#endif
