+++ pvzclass/Events/DrawUITopEvent.cpp
@@ -1,0 +1,14 @@
+#include "DrawUITopEvent.h"
+
+DrawUITopEvent::DrawUITopEvent()
+{
+	address = 0x4172F6;
+}
+
+void DrawUITopEvent::handle(CONTEXT& context)
+{
+	for (int i = 0; i < listeners.size(); i++)
+	{
+		listeners[i]();
+	}
+}
