+++ pvzclass/Events/DrawUITopEvent.h
@@ -11,16 +11,3 @@ public:
 	DrawUITopEvent();
 	void handle(CONTEXT& context) override;
 };
-
-DrawUITopEvent::DrawUITopEvent()
-{
-	address = 0x4172F6;
-}
-
-void DrawUITopEvent::handle(CONTEXT& context)
-{
-	for (int i = 0; i < listeners.size(); i++)
-	{
-		listeners[i]();
-	}
-}
