+++ pvzclass/Events/ZombieButterEvent.cpp
@@ -1,0 +1,26 @@
+#include "ZombieButterEvent.h"
+
+ZombieButterEvent::ZombieButterEvent()
+{
+	address = 0x5326D0;
+}
+
+#if defined(_WIN64)
+void ZombieButterEvent::handle(CONTEXT& context)
+{
+	auto zombie = std::make_shared<PVZ::Zombie>(context.Rax);
+	for (int i = 0; i < listeners.size(); i++)
+	{
+		listeners[i](zombie);
+	}
+}
+#else
+void ZombieButterEvent::handle(CONTEXT& context)
+{
+	auto zombie = std::make_shared<PVZ::Zombie>(context.Eax);
+	for (int i = 0; i < listeners.size(); i++)
+	{
+		listeners[i](zombie);
+	}
+}
+#endif
